Add support for multi-dimensional systems
It was necessary to write a new comparator that is able to deal with multiple dimensions. In the new format, the user has to provide an array with booleans that tell the getParetoFrontier function which dimensions are maximized (true) or minimized (false). The shorthand options topRight, topLeft, bottomRight and bottomLeft can still be used in exactly the same way, in which case, the boolean array is not necessary. Tests for the 3D case were added. As a consequence of this new format, issue #1 was solved.